refactor: domain 패키지 customer 구성
