+++ src/main/java/woowacourse/shoppingcart/domain/customer/PasswordEncryptor.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.domain;
+package woowacourse.shoppingcart.domain.customer;
 
 public interface PasswordEncryptor {
     String encrypt(String password);
