+++ src/main/java/woowacourse/shoppingcart/domain/customer/PhoneNumber.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.domain;
+package woowacourse.shoppingcart.domain.customer;
 
 import java.util.regex.Pattern;
