+++ src/test/java/woowacourse/shoppingcart/acceptance/AcceptanceTest.java
@@ -9,7 +9,6 @@ import org.springframework.test.context.ActiveProfiles;
 import io.restassured.RestAssured;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 @ActiveProfiles("test")
 public class AcceptanceTest {
