+++ src/test/java/woowacourse/shoppingcart/domain/CustomerTest.java
@@ -5,6 +5,8 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import woowacourse.shoppingcart.domain.customer.Customer;
+
 class CustomerTest {
 
     @Test
