+++ src/test/java/woowacourse/shoppingcart/domain/PasswordTest.java
@@ -11,6 +11,7 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
+import woowacourse.shoppingcart.domain.customer.Password;
 import woowacourse.shoppingcart.exception.InvalidFormException;
 import woowacourse.shoppingcart.exception.InvalidLengthException;
