+++ src/test/java/woowacourse/shoppingcart/domain/PhoneNumberTest.java
@@ -10,6 +10,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import woowacourse.shoppingcart.domain.customer.PhoneNumber;
 import woowacourse.shoppingcart.exception.InvalidFormException;
 
 class PhoneNumberTest {
